[WIP] Windows compatible build process
The Build process has been made for linux first. I do think that some features are unix compliant, but some others are not.
I used js to try to unify methods, but since some commands are used, it is complicated.
Maybe the end process is to use JS only methods, such as a js file zipper. We will see