[FIX] Fix shadowing attribute
In VFS (persisted) setup, the parameter name for the vfs path was the same as the attribute used across the codebase. This caused the attribute to be shadowed by the parameter. Since the parameter was "std::move(d)" in the attribute, further usage lead to an empry string. This commit fixes the shadowing by renaming the parameter, and never refer to it again as to use the attribute.